training-optimizer/requirements.txt
# Python 3.10+ required for full compatibility
# If using Python 3.9, garminconnect will use an older version

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6

# Garmin Integration
# Note: garminconnect 0.2.17+ requires Python 3.10+
# For Python 3.9, version 0.2.8 will be installed
garminconnect>=0.2.8

# AI Integration
anthropic==0.18.1

# Database
sqlalchemy==2.0.25
alembic==1.13.1

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
plotly==5.18.0

# Scheduling
apscheduler==3.10.4

# Configuration
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
requests==2.31.0
aiofiles==23.2.1
python-dateutil==2.8.2
pytz==2024.1

# Notifications
emails==0.6

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging
loguru==0.7.2

# Security
cryptography==42.0.0

# Development
black==24.1.1
ruff==0.1.14
